+++ src/componentes/escolas/Paa/ElaboracaoPaa/ElaborarNovoPlano/Relatorios/AtividadesPrevistas/AtividadesPrevistas.js
@@ -1,35 +1,79 @@
-import { useCallback } from "react";
-import { useNavigate } from "react-router-dom";
-import { Spin } from "antd";
-import { RelatorioVisualizacao } from "../components/RelatorioVisualizacao";
-import { useGetPaa } from "../../../../componentes/hooks/useGetPaa";
-import { PaaContext, usePaaContext } from "../../../../componentes/PaaContext";
-import { AtividadesPrevistas } from "./AtividadesPrevistas";
-import { RecursosProprios } from "./RecursosProprios";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { MsgImgCentralizada } from "../../../../../../Globais/Mensagens/MsgImgCentralizada";
+import Img404 from "../../../../../../../assets/img/img-404.svg";
+import { RelatorioTabelaGrupo } from "../components/RelatorioTabelaGrupo";
+import { toastCustom } from "../../../../../../Globais/ToastCustom";
+import { useGetAtividadesEstatutarias } from "./hooks/useGetAtividadesEstatutarias";
+import { useGetAtividadesEstatutariasTabelas } from "./hooks/useGetAtividadesEstatutariasTabelas";
+import { getErrorMessage } from "../../../../../../../utils/obtemMsgErroAxios";
+import { TagRetificacao } from "../../../../componentes/TagRetificacao";
+import {
+  IconButton,
+  EditIconButton,
+} from "../../../../../../Globais/UI/Button";
+import { Icon } from "../../../../../../Globais/UI/Icon";
+import {
+  createAtividadeEstatutariaPaa,
+  updateAtividadeEstatutariaPaa,
+  deleteAtividadeEstatutariaPaa,
+  linkAtividadeEstatutariaExistentePaa,
+} from "../../../../../../../services/escolas/Paa.service";
+import { ModalConfirmarExclusao } from "../../../../../../sme/Parametrizacoes/componentes/ModalConfirmarExclusao";
+import { parseISO, getYear } from "date-fns";
+import { visoesService } from "../../../../../../../services/visoes.service";
 import "./styles.scss";
+import { Spin } from "antd";
 
-export const VisualizarAtividadesPrevistasContent = () => {
-  const navigate = useNavigate();
+const formatarMesAno = (valor) => {
+    if (!valor) return "-";
 
-  const { paa, isFetching: isLoadingPaa } = usePaaContext();
+    const formatData = (data) => {
+        if (Number.isNaN(data.getTime())) return "-";
 
+        const mes = data.toLocaleDateString("pt-BR", { month: "long" });
+        const ano = data.getFullYear();
+        if (!mes) return `${ano || ""}`;
+        return `${mes.charAt(0).toUpperCase()}${mes.slice(1)}/${ano}`;
+    };
 
-  const handleVoltar = useCallback(() => {
-    let voltarRota = "/elaborar-novo-paa";
+    // Parse da string YYYY-MM-DD para evitar problemas de fuso horário
+    const partes = valor.split("-");
 
-    if (paa?.status === "EM_RETIFICACAO") {
-      voltarRota =`/retificacao-paa/${paa?.uuid}`;
-    }
-    navigate(voltarRota, {
-      state: {
-        activeTab: "relatorios",
-        expandedSections: {
-          planoAnual: true,
-          componentes: true,
-        },
-      },
-    });
-  }, [navigate]);
+    if (partes.length !== 3) return formatData(new Date(valor));
+
+    const [anoStr, mesStr, diaStr] = partes;
+    const ano = Number.parseInt(anoStr, 10);
+    const mes = Number.parseInt(mesStr, 10);
+    const dia = Number.parseInt(diaStr, 10);
+
+    if (Number.isNaN(ano) || Number.isNaN(mes) || Number.isNaN(dia)) return "-";
+
+    // Criar data no fuso horário local para evitar problemas de UTC
+    return formatData(new Date(ano, mes - 1, dia));
+};
+
+export const AtividadesPrevistas = () => {
+  const podeEditar = visoesService.getPermissoes(["custom_change_paa"]);
+
+  const [atividadesTabela, setAtividadesTabela] = useState([]);
+  const carregouDadosIniciais = useRef(false);
+  const deveSincronizarTabela = useRef(false);
+  const [isSalvando, setIsSalvando] = useState(false);
+
+  const { atividades, isFetching: isLoading, isError, refetch } = useGetAtividadesEstatutarias();
+  const { data: tabelasAtividades } = useGetAtividadesEstatutariasTabelas();
+  const [modalExcluir, setModalExcluir] = useState({ aberto: false, atividade: null });
+  const [isExcluindoAtividade, setIsExcluindoAtividade] = useState(false);
+
+  const paaUuid = localStorage.getItem("PAA");
+
+  const tiposOptions = useMemo(() => {
+    const tipos = tabelasAtividades?.tipo || [];
+    return tipos.map((item) => ({
+      value: item.key,
+      label: item.value,
+    }));
+  }, [tabelasAtividades]);
 
   useEffect(() => {
     if (isLoading) {
@@ -49,29 +93,6 @@ export const VisualizarAtividadesPrevistasContent = () => {
       deveSincronizarTabela.current = false;
     }
   }, [atividades, isLoading]);
-
-  useEffect(() => {
-    if (isLoadingRecursos) {
-      return;
-    }
-
-    if (!carregouRecursosIniciais.current || deveSincronizarRecursos.current) {
-      const lista = Array.isArray(recursosProprios) ? recursosProprios : [];
-      const recursosFormatados = lista.map((item) => ({
-        ...item,
-        tipoAtividade: item?.fonte_recurso?.nome || "-",
-        data: item?.data_prevista || "",
-        valor: Number(item?.valor) || 0,
-        emEdicao: false,
-        isNovo: false,
-        needsSync: false,
-        _destroy: false,
-      }));
-      setRecursosPropriosTabela(recursosFormatados);
-      carregouRecursosIniciais.current = true;
-      deveSincronizarRecursos.current = false;
-    }
-  }, [recursosProprios, isLoadingRecursos]);
 
   const handleAdicionarAtividade = useCallback(() => {
     if (!podeEditar) return;
@@ -119,18 +140,7 @@ export const VisualizarAtividadesPrevistasContent = () => {
     return mesAnoLabel;
   }, []);
 
-  const handleEditarRecursosProprios = useCallback(() => {
-    navigate("/elaborar-novo-paa?fromAtividadesPrevistas=1", {
-      state: {
-        activeTab: "receitas",
-        receitasDestino: "recursos-proprios",
-        fromRecursosPropriosRelatorio: true,
-      },
-    });
-  }, [navigate]);
-
-  const handleChangeAtividade = useCallback(
-    (uuid, campo, valor, extra = {}) => {
+  const handleChangeAtividade = useCallback((uuid, campo, valor, extra = {}) => {
       if (!podeEditar) return;
       setAtividadesTabela((prev) =>
         prev.map((item) => {
@@ -150,17 +160,10 @@ export const VisualizarAtividadesPrevistasContent = () => {
           }
 
           if (campo === "data") {
-            const dataObj = valor ? new Date(valor) : null;
-            if (dataObj && !Number.isNaN(dataObj.getTime())) {
-              updated.mesAno = formatarMesAno(valor);
-            } else {
-              updated.mesAno = "-";
-            }
-
+            const mesAno = formatarMesAno(valor);
+            updated.mesAno = mesAno;
             if (!item.emEdicao) {
-              updated.needsSync = Boolean(
-                dataObj && !Number.isNaN(dataObj.getTime()),
-              );
+              updated.needsSync = mesAno !== "-";
             }
           }
 
@@ -171,7 +174,7 @@ export const VisualizarAtividadesPrevistasContent = () => {
     [podeEditar],
   );
 
-  const handleEditarAtividade = useCallback(
+  const handleEditar = useCallback(
     (atividade) => {
       if (!podeEditar) return;
       setAtividadesTabela((prev) =>
@@ -217,6 +220,13 @@ export const VisualizarAtividadesPrevistasContent = () => {
     },
     [podeEditar],
   );
+
+  const getAtividadeEstatutariaUuid = (atividade => {
+      return atividade?.atividade_estatutaria?.uuid ||
+            atividade?.atividade_estatutaria ||
+            atividade?.atividadeEstatutariaUuid ||
+            atividade?.uuid;
+  })
 
   const handleSalvarAtividades = useCallback(async () => {
     if (!podeEditar) return;
@@ -266,24 +276,16 @@ export const VisualizarAtividadesPrevistasContent = () => {
     const atividadesPendentes = atividadesPreparadas.filter(
       (item) => item.needsSync,
     );
-    const recursosPendentes = recursosPropriosTabela.filter(
-      (item) => item.needsSync,
-    );
 
-    if (atividadesPendentes.length === 0 && recursosPendentes.length === 0) {
+    if (atividadesPendentes.length === 0) {
       return;
     }
 
     setIsSalvando(true);
 
     try {
-      if (atividadesPendentes.length > 0) {
         for (const atividade of atividadesPendentes) {
-          const atividadeEstatutariaUuid =
-            atividade?.atividade_estatutaria?.uuid ||
-            atividade?.atividade_estatutaria ||
-            atividade?.atividadeEstatutariaUuid ||
-            atividade?.uuid;
+          const atividadeEstatutariaUuid = getAtividadeEstatutariaUuid(atividade);
 
           if (atividade._destroy) {
             if (atividade.isNovo) {
@@ -370,81 +372,22 @@ export const VisualizarAtividadesPrevistasContent = () => {
                 : item,
             ),
         );
-      }
-
-      if (recursosPendentes.length > 0) {
-        for (const recurso of recursosPendentes) {
-          if (recurso._destroy) {
-            if (recurso.isNovo) {
-              continue;
-            }
-            if (!recurso.uuid) {
-              throw new Error(
-                "Identificador do recurso próprio não encontrado.",
-              );
-            }
-            // eslint-disable-next-line no-await-in-loop
-            await deleteRecursoProprioPaa(recurso.uuid);
-            toastCustom.ToastCustomSuccess(
-              "Sucesso!",
-              "Recurso próprio excluído com sucesso.",
-            );
-          }
-        }
-
-        deveSincronizarRecursos.current = true;
-        await refetchRecursos();
-
-        const recursosProcessados = new Set(
-          recursosPendentes.map((item) => item.uuid),
-        );
-        setRecursosPropriosTabela((prev) =>
-          prev
-            .filter(
-              (item) => !(recursosProcessados.has(item.uuid) && item._destroy),
-            )
-            .map((item) =>
-              recursosProcessados.has(item.uuid)
-                ? {
-                    ...item,
-                    needsSync: false,
-                    isNovo: false,
-                    _destroy: false,
-                  }
-                : item,
-            ),
-        );
-      }
     } catch (error) {
-      console.error("Erro ao salvar alterações:", error);
-      const mensagem = getErrorMessage(
-        error,
-        "Não foi possível salvar as alterações. Tente novamente.",
-      );
-      toastCustom.ToastCustomError("Erro!", mensagem);
-    } finally {
-      setIsSalvando(false);
-    }
-  }, [
-    atividadesTabela,
-    recursosPropriosTabela,
-    paaUuid,
-    refetch,
-    refetchRecursos,
-    podeEditar,
-  ]);
-
-  const handleEditarLinha = useCallback(
-    (atividade) => {
-      if (!podeEditar) return;
-      setAtividadesTabela((prev) =>
-        prev.map((item) =>
-          item.uuid === atividade.uuid ? { ...item, emEdicao: true } : item,
-        ),
-      );
-    },
-    [podeEditar],
-  );
+          console.error("Erro ao salvar alterações:", error);
+          const mensagem = getErrorMessage(
+            error,
+            "Não foi possível salvar as alterações. Tente novamente.",
+          );
+          toastCustom.ToastCustomError("Erro!", mensagem);
+        } finally {
+          setIsSalvando(false);
+        }
+    }, [
+        atividadesTabela,
+        paaUuid,
+        refetch,
+        podeEditar,
+    ]);
 
   const handleExcluirAtividade = useCallback(
     (atividade) => {
@@ -472,11 +415,7 @@ export const VisualizarAtividadesPrevistasContent = () => {
           prev.filter((item) => item.uuid !== atividade.uuid),
         );
       } else {
-        const atividadeEstatutariaUuid =
-          atividade?.atividade_estatutaria?.uuid ||
-          atividade?.atividade_estatutaria ||
-          atividade?.atividadeEstatutariaUuid ||
-          atividade?.uuid;
+        const atividadeEstatutariaUuid = getAtividadeEstatutariaUuid(atividade);
 
         if (!atividadeEstatutariaUuid) {
           throw new Error("Identificador da atividade não encontrado.");
@@ -515,116 +454,15 @@ export const VisualizarAtividadesPrevistasContent = () => {
     setModalExcluir({ aberto: false, atividade: null });
   }, []);
 
-  const handleExcluirRecursoProprio = useCallback(
-    (recurso) => {
-      if (!podeEditar) return;
-      setModalExcluirRecurso({ aberto: true, recurso });
-    },
-    [podeEditar],
-  );
-
-  const confirmarExclusaoRecursoProprio = useCallback(async () => {
-    if (!podeEditar || isExcluindoRecurso) {
-      return;
-    }
-
-    const recurso = modalExcluirRecurso.recurso;
-    if (!recurso) {
-      return;
-    }
-
-    try {
-      setIsExcluindoRecurso(true);
-
-      if (recurso.isNovo) {
-        setRecursosPropriosTabela((prev) =>
-          prev.filter((item) => item.uuid !== recurso.uuid),
-        );
-        toastCustom.ToastCustomSuccess(
-          "Sucesso!",
-          "Recurso próprio excluído com sucesso.",
-        );
-      } else {
-        if (!recurso.uuid) {
-          throw new Error("Identificador do recurso próprio não encontrado.");
-        }
-
-        await deleteRecursoProprioPaa(recurso.uuid);
-        toastCustom.ToastCustomSuccess(
-          "Sucesso!",
-          "Recurso próprio excluído com sucesso.",
-        );
-
-        deveSincronizarRecursos.current = true;
-        await refetchRecursos();
-
-        setRecursosPropriosTabela((prev) =>
-          prev.filter((item) => item.uuid !== recurso.uuid),
-        );
-      }
-    } catch (error) {
-      console.error("Erro ao excluir recurso próprio:", error);
-      toastCustom.ToastCustomError(
-        "Erro!",
-        "Não foi possível excluir o recurso próprio. Tente novamente.",
-      );
-    } finally {
-      setIsExcluindoRecurso(false);
-      setModalExcluirRecurso({ aberto: false, recurso: null });
-    }
-  }, [
-    isExcluindoRecurso,
-    modalExcluirRecurso.recurso,
-    refetchRecursos,
-    podeEditar,
-  ]);
-
-  const cancelarExclusaoRecursoProprio = useCallback(() => {
-    if (isExcluindoRecurso) {
-      return;
-    }
-
-    setModalExcluirRecurso({ aberto: false, recurso: null });
-  }, [isExcluindoRecurso]);
-
-  const formatarData = (valor) => {
-    if (!valor) return "-";
-    const date = new Date(valor);
-    return Number.isNaN(date.getTime())
-      ? "-"
-      : new Intl.DateTimeFormat("pt-BR").format(date);
-  };
-
-  const formatarValor = (valor) =>
-    new Intl.NumberFormat("pt-BR", {
-      minimumFractionDigits: 2,
-      maximumFractionDigits: 2,
-    }).format(Number(valor) || 0);
 
   const atividadesVisiveis = useMemo(
     () => atividadesTabela.filter((item) => !item._destroy),
     [atividadesTabela],
   );
 
-  const recursosPropriosVisiveis = useMemo(
-    () => recursosPropriosTabela.filter((item) => !item._destroy),
-    [recursosPropriosTabela],
-  );
-
-  const totalRecursosProprios = useMemo(
-    () =>
-      recursosPropriosVisiveis.reduce(
-        (acc, item) => acc + (Number(item.valor) || 0),
-        0,
-      ),
-    [recursosPropriosVisiveis],
-  );
-
   const temAlteracaoPendente = useMemo(
-    () =>
-      atividadesTabela.some((item) => item.needsSync) ||
-      recursosPropriosTabela.some((item) => item.needsSync),
-    [atividadesTabela, recursosPropriosTabela],
+    () => atividadesTabela.some((item) => item.needsSync),
+    [atividadesTabela],
   );
 
   const colunas = useMemo(
@@ -661,7 +499,10 @@ export const VisualizarAtividadesPrevistasContent = () => {
               ))}
             </select>
           ) : (
-            record.tipoAtividade || "-"
+            <>
+            {record.tipoAtividade || "-"}
+            <div>{record?.alteracao && <TagRetificacao />}</div>
+            </>
           ),
       },
       {
@@ -805,9 +646,7 @@ export const VisualizarAtividadesPrevistasContent = () => {
                       <EditIconButton
                         className="p-0"
                         onClick={() =>
-                          record.isNovo
-                            ? handleEditarLinha(record)
-                            : handleEditarAtividade(record)
+                          record.isNovo ? handleEditar(record): null
                         }
                       />
                       <IconButton
@@ -831,67 +670,15 @@ export const VisualizarAtividadesPrevistasContent = () => {
       },
     ],
     [
-      handleChangeAtividade,
-      handleEditarAtividade,
-      handleEditarLinha,
-      handleExcluirAtividade,
-      handleSalvarLinha,
-      tiposOptions,
-      isSalvando,
-      podeEditar,
+        obterMesAno,
+        handleChangeAtividade,
+        handleEditar,
+        handleExcluirAtividade,
+        handleSalvarLinha,
+        tiposOptions,
+        isSalvando,
+        podeEditar,
     ],
-  );
-
-  const recursosColumns = useMemo(
-    () => [
-      {
-        title: "Tipo de Atividades",
-        dataIndex: "tipoAtividade",
-        key: "tipoAtividade",
-        render: (text) => text || "-",
-      },
-      {
-        title: "Data",
-        dataIndex: "data",
-        key: "data",
-        render: (texto) => formatarData(texto),
-        width: 150,
-      },
-      {
-        title: "Descrição das Receitas e Atividades Previstas",
-        dataIndex: "descricao",
-        key: "descricao",
-        render: (text) => text || "-",
-      },
-      {
-        title: "Valor Estimado",
-        dataIndex: "valor",
-        key: "valor",
-        align: "right",
-        render: (valor) => formatarValor(valor),
-      },
-      {
-        title: "Ações",
-        key: "acoes",
-        width: 80,
-        render: (_, record) =>
-          podeEditar ? (
-            <div className="atividades-previstas__acoes-recursos">
-              <IconButton
-                className="p-0"
-                icon="faTrashCan"
-                tooltipMessage="Excluir"
-                iconProps={{
-                  style: { color: "#00585E" },
-                }}
-                aria-label="Excluir recurso próprio"
-                onClick={() => handleExcluirRecursoProprio(record)}
-              />
-            </div>
-          ) : null,
-      },
-    ],
-    [handleExcluirRecursoProprio, podeEditar],
   );
 
   const botoesHeader = (
@@ -920,119 +707,46 @@ export const VisualizarAtividadesPrevistasContent = () => {
 
   const shouldShowEmpty =
     !isLoading &&
-    !isLoadingRecursos &&
     !isError &&
     carregouDadosIniciais.current &&
-    carregouRecursosIniciais.current &&
-    atividadesVisiveis.length === 0 &&
-    recursosPropriosVisiveis.length === 0;
+    atividadesVisiveis.length === 0
 
   return (
     <>
-      <RelatorioVisualizacao
-        title="Atividades previstas"
-        onBack={handleVoltar}
-        isLoading={isLoading || isLoadingRecursos}
-        error={isError}
-        errorContent={
-          <MsgImgCentralizada
-            texto="Não foi possível carregar as atividades estatutárias."
-            img={Img404}
-            dataQa="atividades-previstas-erro"
-          />
-        }
-        isEmpty={shouldShowEmpty}
-        emptyContent={
-          <MsgImgCentralizada
-            texto="Utilize o registro de prioridades para manter as atividades previstas do PAA atualizadas."
-            img={Img404}
-            dataQa="atividades-previstas-sem-dados"
-          />
-        }
-        showWatermark={true}
-        heightDeps={[
-          atividadesTabela,
-          recursosPropriosTabela,
-          isLoading,
-          isLoadingRecursos,
-          isError,
-        ]}
-      >
-        {!shouldShowEmpty && !isError && (
-          <RelatorioTabelaGrupo
-            title="Atividades Estatutárias"
-            columns={colunas}
-            dataSource={atividadesVisiveis}
-            rowKey={(item) => item.uuid}
-            headerExtra={botoesHeader}
-            containerClassName="relatorio-tabela-grupo atividades-previstas__tabela"
-          />
-        )}
-        <RelatorioTabelaGrupo
-          title="Recursos próprios"
-          columns={recursosColumns}
-          dataSource={recursosPropriosVisiveis}
-          rowKey={(item) => item.uuid}
-          headerExtra={
-            <button
-              type="button"
-              className="atividades-previstas__btn-editar-recursos"
-              onClick={handleEditarRecursosProprios}
-            >
-              <Icon
-                icon="faEdit"
-                iconProps={{ style: { marginRight: 8, color: "#FFFFFF" } }}
-              />
-              Editar receitas de recursos próprios
-            </button>
-          }
-          containerClassName="relatorio-tabela-grupo atividades-previstas__tabela mt-4"
-          tableProps={{
-            summary: () => (
-              <Table.Summary.Row className="atividades-previstas__tabela-resumo">
-                <Table.Summary.Cell index={0}>
-                  <span className="atividades-previstas__resumo-label">
-                    Tipo de Atividades
-                  </span>
-                </Table.Summary.Cell>
-                <Table.Summary.Cell index={1} />
-                <Table.Summary.Cell index={2} />
-                <Table.Summary.Cell index={3} align="right">
-                  <span className="atividades-previstas__resumo-valor">
-                    {formatarValor(totalRecursosProprios)}
-                  </span>
-                </Table.Summary.Cell>
-                <Table.Summary.Cell index={4} />
-              </Table.Summary.Row>
-            ),
-          }}
-        />
-      </RelatorioVisualizacao>
+        <Spin spinning={isLoading || isSalvando} size="large">
+            {isError && <MsgImgCentralizada
+                texto="Não foi possível carregar as atividades estatutárias."
+                img={Img404}
+                dataQa="atividades-previstas-erro"
+            />}
+            {shouldShowEmpty && <MsgImgCentralizada
+                texto="Utilize o registro de prioridades para manter as atividades previstas do PAA atualizadas."
+                img={Img404}
+                dataQa="atividades-previstas-sem-dados"
+            />}
+            {!shouldShowEmpty && !isError && (
+            <RelatorioTabelaGrupo
+                title="Atividades Estatutárias"
+                columns={colunas}
+                dataSource={atividadesVisiveis}
+                rowKey={(item) => item.uuid}
+                headerExtra={botoesHeader}
+                containerClassName="relatorio-tabela-grupo atividades-previstas__tabela"
+            />
+            )}
+        </Spin>
 
-      <ModalConfirmarExclusao
-        open={modalExcluir.aberto}
-        onOk={confirmarExclusaoAtividade}
-        okText="Excluir"
-        okButtonProps="btn-base-vermelho"
-        onCancel={cancelarExclusaoAtividade}
-        cancelText="Voltar"
-        cancelButtonProps="btn-base-verde-outline"
-        titulo="Excluir atividade"
-        bodyText="Tem certeza que deseja excluir a atividade selecionada?"
-      />
-      <ModalConfirmarExclusao
-        open={modalExcluirRecurso.aberto}
-        onOk={confirmarExclusaoRecursoProprio}
-        okText="Excluir"
-        okButtonProps="btn-base-vermelho"
-        onCancel={cancelarExclusaoRecursoProprio}
-        cancelText="Voltar"
-        cancelButtonProps="btn-base-verde-outline"
-        titulo="Excluir recurso próprio"
-        bodyText="Tem certeza que deseja excluir o recurso próprio selecionado?"
-      />
+        <ModalConfirmarExclusao
+            open={modalExcluir.aberto}
+            onOk={confirmarExclusaoAtividade}
+            okText="Excluir"
+            okButtonProps="btn-base-vermelho"
+            onCancel={cancelarExclusaoAtividade}
+            cancelText="Voltar"
+            cancelButtonProps="btn-base-verde-outline"
+            titulo="Excluir atividade"
+            bodyText="Tem certeza que deseja excluir a atividade selecionada?"
+        />
     </>
   );
 };
-
-export default VisualizarAtividadesPrevistas;
