+++ src/componentes/escolas/Paa/componentes/TagRetificacao/index.js
@@ -1,12 +1,12 @@
 import { Tag } from "antd";
 
-export const TagRetificacao = () => {
+export const TagRetificacao = ({label}) => {
     const styleTag = {
         backgroundColor: 'white',
         color: '#01585E',
         border: '2px solid #01585E'
     }
     return (
-        <Tag style={styleTag}>Em retificação</Tag>
+        <Tag style={styleTag}>{label || 'Em retificação'}</Tag>
     )
 }
